+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### HEAD
 
 ```
+* Add Dependency Status
 * Add Build Status for Win
 * Add appveyor.yml
 ```